first stab at building the worldwind.jar with SBT, to (1) enable git project dependencies in other SBT projects, and (2) use SBT's artifact publishing infrastructure
removed the jogl and gluegen jars, which can be resolved through SBT.
gdal.jar still has to stay in there as unmanaged lib since it is NOT
identical with the Maven central artifact - there are additional classes
in the local version WorldWindJava depends on. Fix this

The general approach was to leave files/directories in the same place,
to simplify merges from the WorldWindJava upstream repo. This causes
considerable SBT overhead since the current layout is not very
Maven/Gradle/SBT compatible.